back-end/requirements.txt
asgiref==3.6.0
coverage==7.2.3
ddt==1.6.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
iniconfig==2.0.0
Markdown==3.4.3
mysqlclient==2.1.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
PyJWT==2.7.0
pytest==7.3.1
pytest-cov==4.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
